requirements.txt
virtualenv
python-dotenv 
ipython
pyaudio
openai
pygame
google-generativeai
SpeechRecognition
undetected_chromedriver
selenium